# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GymPal Backend API",
    "description": "Complete API for the GymPal fitness application with social features, personal tracking, workouts, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "GymPal Team",
      "email": "support@gympal.app",
      "url": "https://gympal.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.gympal.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "description": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "SecurePass123!",
                "username": "fitnesslover",
                "full_name": "John Doe",
                "date_of_birth": "1995-06-15",
                "gender": "male",
                "terms_accepted": true,
                "privacy_policy_accepted": true,
                "bio": "Fitness enthusiast and personal trainer",
                "fitness_level": "beginner",
                "timezone": "UTC",
                "language": "en"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user and get access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "SecurePass123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "user": {
                      "id": "03f0af5a-8976-4f5c-8893-d1a16ce45a11",
                      "email": "user@example.com",
                      "username": "fitnesslover",
                      "fullName": "John Doe",
                      "emailVerified": true,
                      "createdAt": "2024-01-15T10:30:00Z",
                      "updatedAt": "2024-01-20T14:45:00Z"
                    },
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIwM2YwYWY1YS04OTc2LTRmNWMtODg5My1kMWExNmNlNDVhMTEiLCJleHAiOjE3NjE0MjAwODh9.placeholder"
                  },
                  "message": "Login successful"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "description": "Get current authenticated user information",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User information retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User logout",
        "description": "Logout current user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logout successful"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh token",
        "description": "Refresh access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIwM2YwYWY1YS04OTc2LTRmNWMtODg5My1kMWExNmNlNDVhMTEiLCJleHAiOjE3NjE0MjAwODh9.placeholder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed"
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Reset password",
        "description": "Reset password using reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "new_password"],
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "new_password": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset email sent"
          }
        }
      }
    },
    "/api/v1/auth/change-password/{id}": {
      "put": {
        "tags": ["Authentication"],
        "summary": "Change password",
        "description": "Change user password",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currentPassword", "newPassword"],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          }
        }
      }
    },
    "/api/v1/auth/delete-account/{id}": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "Delete account",
        "description": "Delete user account permanently",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "03f0af5a-8976-4f5c-8893-d1a16ce45a11"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "fitnesslover"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "emailVerified": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-20T14:45:00Z"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "username", "full_name", "date_of_birth", "gender", "terms_accepted", "privacy_policy_accepted"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 100,
            "example": "SecurePass123!"
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "fitnesslover"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "John Doe"
          },
          "date_of_birth": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "description": "Date in YYYY-MM-DD format",
            "example": "1995-06-15"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "other", "prefer_not_to_say"],
            "example": "male"
          },
          "terms_accepted": {
            "type": "boolean",
            "description": "Must be true to accept terms",
            "example": true
          },
          "privacy_policy_accepted": {
            "type": "boolean",
            "description": "Must be true to accept privacy policy",
            "example": true
          },
          "bio": {
            "type": "string",
            "maxLength": 500,
            "example": "Fitness enthusiast and personal trainer"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/avatar.jpg"
          },
          "fitness_level": {
            "type": "string",
            "enum": ["beginner", "intermediate", "advanced", "expert"],
            "example": "beginner"
          },
          "timezone": {
            "type": "string",
            "example": "UTC"
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5,
            "example": "en"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string"
          },
          "expiresIn": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          }
        }
      },
...
    }
  }
}